consensus: Move BlockVerifier and tests into block.rs
Move BlockVerifier and its tests from src/verify.rs to
src/verify/block.rs.

Part of #428.
